SDK-376: Revert the refactoring of cluster_cmd_line (#286)
Revert the refactoring of cluster_cmd_line
